services/pdf-service/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.1
pydantic>=2.7.4
reportlab>=4.2.0
boto3>=1.34.120
python-dotenv>=1.0.1
